Moved the mapViewDelegate to it's own class
Move the mapview delegate to it's own class so that it can be reused on both View controllers
